Promisify tasks and factor to support build.js. Add scripts to package.json. Update .travis.yml
