+++ .travis.yml
@@ -3,7 +3,7 @@ sudo: false
 node_js:
 - '6'
 before_script:
-- node tasks/webpack.js
+- node tasks/build.js
 script:
 - node tasks/saucetests.js
 addons:
