+++ package.json
@@ -25,5 +25,11 @@
     "sauce-test-runner": "^0.1.7",
     "sinon": "^2.0.0-pre.5",
     "webpack": "^1.14.0"
+  },
+  "scripts": {
+    "build": "node tasks/build.js",
+    "watch": "node tasks/watch.js",
+    "lint": "node tasks/lint.js",
+    "saucetests": "node tasks/saucetests.js"
   }
 }
