+++ tasks/lint.js
@@ -1,31 +1,6 @@
 /*jslint node: true */
 'use strict';
 
-const spawn = require('child_process').spawn;
-const glob = require('glob');
+const lint = require('./lintHelper');
 
-const lintFiles = [
-      'gulpfile.js',
-      'elements/*.js',
-      'mixins/*.js',
-      'test/**/*.js',
-      'tasks/*.js'
-  ];
-
-//
-// Runs jshint
-//
-function runLint() {
-  let args = [];
-  lintFiles.forEach(item => {
-    let globList = glob.sync(item);
-    args = args.concat(globList);
-  });
-
-  spawn(
-    'jshint', 
-    args,
-    {stdio: 'inherit'});
-}
-
-runLint();
+lint();
